range muss Double akzeptieren

+++ src/main/java/de/nadu_ocholt/anbauplaner/application/plant/RangeCmDTO.java
@@ -1,4 +1,4 @@
 package de.nadu_ocholt.anbauplaner.application.plant;
 
-public record RangeCmDTO(Integer min, Integer max) {
+public record RangeCmDTO(Double min, Double max) {
 }
